+++ app/(tabs)/index.tsx
@@ -1,98 +1,292 @@
-import { Image } from 'expo-image';
-import { Platform, StyleSheet } from 'react-native';
-
-import { HelloWave } from '@/components/hello-wave';
-import ParallaxScrollView from '@/components/parallax-scroll-view';
-import { ThemedText } from '@/components/themed-text';
-import { ThemedView } from '@/components/themed-view';
-import { Link } from 'expo-router';
+import { sampleCourse } from '@/src/data/sampleCourse';
+import { useAppStore } from '@/src/store';
+import { useRouter } from 'expo-router';
+import React, { useEffect } from 'react';
+import { ActivityIndicator, Alert, ScrollView, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
 
 export default function HomeScreen() {
-  return (
-    <ParallaxScrollView
-      headerBackgroundColor={{ light: '#A1CEDC', dark: '#1D3D47' }}
-      headerImage={
-        <Image
-          source={require('@/assets/images/partial-react-logo.png')}
-          style={styles.reactLogo}
-        />
-      }>
-      <ThemedView style={styles.titleContainer}>
-        <ThemedText type="title">Welcome!</ThemedText>
-        <HelloWave />
-      </ThemedView>
-      <ThemedView style={styles.stepContainer}>
-        <ThemedText type="subtitle">Step 1: Try it</ThemedText>
-        <ThemedText>
-          Edit <ThemedText type="defaultSemiBold">app/(tabs)/index.tsx</ThemedText> to see changes.
-          Press{' '}
-          <ThemedText type="defaultSemiBold">
-            {Platform.select({
-              ios: 'cmd + d',
-              android: 'cmd + m',
-              web: 'F12',
-            })}
-          </ThemedText>{' '}
-          to open developer tools.
-        </ThemedText>
-      </ThemedView>
-      <ThemedView style={styles.stepContainer}>
-        <Link href="/modal">
-          <Link.Trigger>
-            <ThemedText type="subtitle">Step 2: Explore</ThemedText>
-          </Link.Trigger>
-          <Link.Preview />
-          <Link.Menu>
-            <Link.MenuAction title="Action" icon="cube" onPress={() => alert('Action pressed')} />
-            <Link.MenuAction
-              title="Share"
-              icon="square.and.arrow.up"
-              onPress={() => alert('Share pressed')}
-            />
-            <Link.Menu title="More" icon="ellipsis">
-              <Link.MenuAction
-                title="Delete"
-                icon="trash"
-                destructive
-                onPress={() => alert('Delete pressed')}
-              />
-            </Link.Menu>
-          </Link.Menu>
-        </Link>
+    const router = useRouter();
+    const { isLoading, setLoading } = useAppStore();
 
-        <ThemedText>
-          {`Tap the Explore tab to learn more about what's included in this starter app.`}
-        </ThemedText>
-      </ThemedView>
-      <ThemedView style={styles.stepContainer}>
-        <ThemedText type="subtitle">Step 3: Get a fresh start</ThemedText>
-        <ThemedText>
-          {`When you're ready, run `}
-          <ThemedText type="defaultSemiBold">npm run reset-project</ThemedText> to get a fresh{' '}
-          <ThemedText type="defaultSemiBold">app</ThemedText> directory. This will move the current{' '}
-          <ThemedText type="defaultSemiBold">app</ThemedText> to{' '}
-          <ThemedText type="defaultSemiBold">app-example</ThemedText>.
-        </ThemedText>
-      </ThemedView>
-    </ParallaxScrollView>
-  );
+    const fetchHomeData = () => {
+        setLoading(true);
+        
+        const timeoutId = setTimeout(() => {
+            Alert.alert(
+                'Request Timeout',
+                'The server took too long to respond. Please check your connection.',
+                [
+                    { text: 'Retry', onPress: fetchHomeData },
+                    { text: 'Cancel', onPress: () => setLoading(false), style: 'cancel' }
+                ]
+            );
+        }, 10000);
+
+        const successId = setTimeout(() => {
+            clearTimeout(timeoutId);
+            setLoading(false);
+        }, 1500);
+
+        return () => {
+            clearTimeout(timeoutId);
+            clearTimeout(successId);
+        };
+    };
+
+    useEffect(() => {
+        const cleanup = fetchHomeData();
+        return cleanup;
+    }, []);
+
+    if (isLoading) {
+        return (
+            <View style={styles.skeletonContainer}>
+                <ActivityIndicator size="large" color="#19c3e6" />
+                <View style={styles.skeletonTextContainer}>
+                    <View style={styles.skeletonTitle} />
+                    <View style={styles.skeletonSubtitle} />
+                </View>
+                <View style={styles.skeletonCard} />
+                <View style={styles.skeletonCard} />
+            </View>
+        );
+    }
+
+    return (
+        <ScrollView 
+            className="flex-1 bg-gray-50 dark:bg-slate-800"
+            contentContainerStyle={styles.scrollContent}
+            showsVerticalScrollIndicator={false}
+        >
+            {/* Header */}
+            <View style={styles.headerSection}>
+                <Text style={styles.headerIcon}>? </Text>
+                <Text style={styles.title}>Welcome to TeachLink</Text>
+                <Text style={styles.subtitle}>Share and consume knowledge on the go</Text>
+            </View>
+
+            {/* Course Viewer Button - Primary */}
+            <TouchableOpacity
+                style={styles.primaryButton}
+                onPress={() =>
+                    router.push({
+                        pathname: '/course-viewer',
+                        params: { course: JSON.stringify(sampleCourse) }
+                    })
+                }
+                accessibilityRole="button"
+                accessibilityLabel="Start Learning"
+                accessibilityHint="Opens the course viewer with a sample lesson"
+            >
+                <View style={styles.buttonContent}>
+                    <Text style={styles.buttonIcon}>? </Text>
+                    <View style={styles.buttonTextContainer}>
+                        <Text style={styles.buttonTitle}>Start Learning</Text>
+                        <Text style={styles.buttonSubtitle}>Open course viewer</Text>
+                    </View>
+                </View>
+            </TouchableOpacity>
+
+            {/* Search Button */}
+            <TouchableOpacity
+                style={styles.secondaryButton}
+                onPress={() => router.push('/search')}
+                accessibilityRole="button"
+                accessibilityLabel="Search courses"
+                accessibilityHint="Navigates to the search screen"
+            >
+                <View style={styles.secondaryButtonContent}>
+                    <Text style={styles.secondaryIcon}>? </Text>
+                    <View style={styles.secondaryTextContainer}>
+                        <Text style={styles.secondaryTitle}>Search</Text>
+                        <Text style={styles.secondarySubtitle}>Find courses and lessons</Text>
+                    </View>
+                    <Text style={styles.arrow}>{'>'}</Text>
+                </View>
+            </TouchableOpacity>
+
+            {/* Secondary Buttons */}
+            <View style={styles.secondaryButtonsContainer}>
+                {/* Profile Button */}
+                <TouchableOpacity
+                    style={styles.secondaryButton}
+                    onPress={() => router.push('/profile/123')}
+                    accessibilityRole="button"
+                    accessibilityLabel="My Profile"
+                    accessibilityHint="View your learning progress and achievements"
+                >
+                    <View style={styles.secondaryButtonContent}>
+                        <Text style={styles.secondaryIcon}>? </Text>
+                        <View style={styles.secondaryTextContainer}>
+                            <Text style={styles.secondaryTitle}>My Profile</Text>
+                            <Text style={styles.secondarySubtitle}>View your progress</Text>
+                        </View>
+                        <Text style={styles.arrow}>{'>'}</Text>
+                    </View>
+                </TouchableOpacity>
+
+                {/* Settings Button */}
+                <TouchableOpacity
+                    style={styles.secondaryButton}
+                    onPress={() => router.push('/settings')}
+                    accessibilityRole="button"
+                    accessibilityLabel="Settings"
+                    accessibilityHint="Customize your app experience"
+                >
+                    <View style={styles.secondaryButtonContent}>
+                        <Text style={styles.secondaryIcon}>? </Text>
+                        <View style={styles.secondaryTextContainer}>
+                            <Text style={styles.secondaryTitle}>Settings</Text>
+                            <Text style={styles.secondarySubtitle}>Customize your experience</Text>
+                        </View>
+                        <Text style={styles.arrow}>{'>'}</Text>
+                    </View>
+                </TouchableOpacity>
+            </View>
+
+          
+        </ScrollView>
+    );
 }
 
 const styles = StyleSheet.create({
-  titleContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 8,
-  },
-  stepContainer: {
-    gap: 8,
-    marginBottom: 8,
-  },
-  reactLogo: {
-    height: 178,
-    width: 290,
-    bottom: 0,
-    left: 0,
-    position: 'absolute',
-  },
+    skeletonContainer: {
+        flex: 1,
+        padding: 24,
+        alignItems: 'center',
+        backgroundColor: '#F9FAFB',
+        marginTop: 40,
+    },
+    skeletonTextContainer: {
+        alignItems: 'center',
+        marginTop: 20,
+        marginBottom: 30,
+        width: '100%',
+    },
+    skeletonTitle: {
+        width: '70%',
+        height: 28,
+        backgroundColor: '#E5E7EB',
+        borderRadius: 8,
+        marginBottom: 12,
+    },
+    skeletonSubtitle: {
+        width: '50%',
+        height: 16,
+        backgroundColor: '#E5E7EB',
+        borderRadius: 4,
+    },
+    skeletonCard: {
+        width: '100%',
+        height: 80,
+        backgroundColor: '#E5E7EB',
+        borderRadius: 12,
+        marginBottom: 16,
+    },
+    scrollContent: {
+        flexGrow: 1,
+        paddingBottom: 30,
+    },
+    headerSection: {
+        paddingHorizontal: 24,
+        paddingVertical: 32,
+        alignItems: 'center',
+    },
+    headerIcon: {
+        fontSize: 60,
+        marginBottom: 16,
+    },
+    title: {
+        fontSize: 28,
+        fontWeight: 'bold',
+        color: '#111827',
+        marginBottom: 8,
+        textAlign: 'center',
+    },
+    subtitle: {
+        fontSize: 16,
+        color: '#6b7280',
+        textAlign: 'center',
+        fontWeight: '500',
+    },
+    primaryButton: {
+        marginHorizontal: 16,
+        marginBottom: 20,
+        backgroundColor: '#19c3e6',
+        borderRadius: 12,
+        paddingHorizontal: 20,
+        paddingVertical: 16,
+        shadowColor: '#000',
+        shadowOffset: { width: 0, height: 2 },
+        shadowOpacity: 0.1,
+        shadowRadius: 4,
+        elevation: 3,
+    },
+    buttonContent: {
+        flexDirection: 'row',
+        alignItems: 'center',
+    },
+    buttonIcon: {
+        fontSize: 28,
+        marginRight: 12,
+    },
+    buttonTextContainer: {
+        flex: 1,
+    },
+    buttonTitle: {
+        fontSize: 16,
+        fontWeight: '700',
+        color: '#ffffff',
+        marginBottom: 2,
+    },
+    buttonSubtitle: {
+        fontSize: 13,
+        color: 'rgba(255, 255, 255, 0.9)',
+        fontWeight: '500',
+    },
+    secondaryButtonsContainer: {
+        marginHorizontal: 16,
+        marginBottom: 24,
+        gap: 12,
+    },
+    secondaryButton: {
+        backgroundColor: '#ffffff',
+        borderRadius: 12,
+        paddingHorizontal: 16,
+        paddingVertical: 14,
+        borderWidth: 1,
+        borderColor: '#e5e7eb',
+        shadowColor: '#000',
+        shadowOffset: { width: 0, height: 1 },
+        shadowOpacity: 0.05,
+        shadowRadius: 2,
+        elevation: 1,
+    },
+    secondaryButtonContent: {
+        flexDirection: 'row',
+        alignItems: 'center',
+    },
+    secondaryIcon: {
+        fontSize: 32,
+        marginRight: 12,
+    },
+    secondaryTextContainer: {
+        flex: 1,
+    },
+    secondaryTitle: {
+        fontSize: 16,
+        fontWeight: '600',
+        color: '#111827',
+        marginBottom: 2,
+    },
+    secondarySubtitle: {
+        fontSize: 13,
+        color: '#6b7280',
+        fontWeight: '500',
+    },
+    arrow: {
+        fontSize: 24,
+        color: '#d1d5db',
+        marginLeft: 8,
+    },
 });
