+++ app/settings.tsx
@@ -1,6 +1,6 @@
+import { useAppStore } from '@/src/store';
 import React from 'react';
-import { View, Text, Switch } from 'react-native';
-import { useAppStore } from '../store';
+import { Switch, Text, View } from 'react-native';
 
 export default function SettingsScreen() {
     const { theme, setTheme } = useAppStore();
